Separate types, utils from WebSocket listener impl
